⭐ feat: Add Upgrade&Destroy Logic

+++ Assets/1. Scripts/Towers/Projectile/Interface/AttackStrategy/AttackStrategy_AquamarineTower.cs
@@ -26,7 +26,7 @@ public class AttackStrategy_AquamarineTower : IAttackStrategy
             if (monster.HasDebuff(DEBUFF_TYPE.Frostbite)) continue;
             // if (monster.HasDebuff(DEBUFF_TYPE.Frostbite) && tower.statHandler.BuildType == TOWER_BUILD_TYPE.Base) continue;
 
-            var debuffObject = PoolManager.Instance.GetFromPool(tower.statHandler.DebuffID, enemy.transform.position, enemy.transform);
+            var debuffObject = PoolManager.Instance.GetFromPool((int)tower.statHandler.DebuffID, enemy.transform.position, enemy.transform);
             if (!debuffObject.TryGetComponent(out BaseDebuff debuff)) continue;
 
             debuff.Apply(monster);
